25/11/21/0020

+++ 3DLv3/Project/BaseProject/BaseProject/src/BaseSystem/CCharaBase.cpp
@@ -14,13 +14,13 @@ CCharaBase::~CCharaBase()
 }
 
 // Å‘åHP‚ğæ“¾
-int CCharaBase::GetMaxHp() const
+float CCharaBase::GetMaxHp() const
 {
 	return mMaxHp;
 }
 
 // Œ»İHP‚ğæ“¾
-int CCharaBase::GetHp() const
+float CCharaBase::GetHp() const
 {
 	return mHp;
 }
