+++ 3DLv3/Project/BaseProject/BaseProject/src/BaseSystem/CCharaBase.h
@@ -14,9 +14,9 @@ public:
 	virtual ~CCharaBase();
 
 	// Å‘åHP‚ğæ“¾
-	int GetMaxHp() const;
+	float GetMaxHp() const;
 	// Œ»İHP‚ğæ“¾
-	int GetHp() const;
+	float GetHp() const;
 
 	// ƒ_ƒ[ƒW‚ğó‚¯‚é
 	virtual void TakeDamage(int damage, CObjectBase* causer);
